POC for the integration tests with the default testing framework
